backend/requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
scikit-learn==1.3.2
joblib==1.3.2
numpy
scipy
tensorflow-cpu>=2.16.1
Pillow==10.2.0
librosa==0.10.1
soundfile==0.12.1
pytest==7.4.3
